pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "engpapersumm"
version = "0.1.0"
description = "Engineering Paper Summarizer Library - Generate application-focused, hierarchical summaries of research papers"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
keywords = ["research", "papers", "summarization", "engineering", "llm", "openai"]
dependencies = [
    "PyPDF2>=3.0.0",
    "scikit-learn>=1.0.0",
    "spacy>=3.0.0",
    "openai>=1.0.0",
    "reportlab>=3.6.0",
    "python-dotenv>=0.19.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "isort>=5.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/engineering-paper-summarizer"
"Bug Tracker" = "https://github.com/yourusername/engineering-paper-summarizer/issues"

[project.scripts]
engpapersumm = "engpapersumm.__main__:main"